% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Import}
\usage{
export(object, con, format, ...)
}
\description{
The function \code{export} save objects from particular file formats. The
rtracklayer package implements support for a number of annotation and
sequence formats.
}
