% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocFile.R
\docType{class}
\name{BiocFile-class}
\alias{BiocFile-class}
\title{Class for High-level File abstraction}
\description{
A \code{RTLFile} object is the base class for classes representing
files accessible with rtracklayer. It wraps a resource (either a path,
URL or connection). We can represent a list of \code{RTLFile} objects
with a \code{RTLFileList}.
}
\author{
Michael Lawrence
}
